html/semantics/embedded-content-0/the-video-element/readyState_during_loadedmetadata.html

<!doctype html>
<html>
 <head>
  <title>video events - readyState property during loadedmetadata</title>
  <script src="/resources/testharness.js"></script>
  <script /resources/testharnessreport.js></script>
  <script src="../../../../common/media.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#mediaevents">spec reference</a></p>
  <video id="v" autoplay controls>
  </video>
  <div id="log"></div>
  <script>
var t = async_test("video.readyState should be >= HAVE_METADATA during loadedmetadata event", {timeout:30000});
var v = document.getElementById("v");
v.addEventListener("loadedmetadata", function() {
  t.step(function() {
   assert_true(v.readyState >= v.HAVE_METADATA);
  });
  t.done();
  v.pause();
}, false);
v.src = getVideoURI("/media/movie_300") + "?" + new Date() + Math.random();
  </script>
 </body>
</html>
